Added a Different Grenade
Added a different Grenade just in blueprints again, this one should work better but need to fix the "throwing" of the grenade as it is using the camera which is third person, so need to change it so it uses the hand of the character or atleast the gun like the grenade launcher already.

Next started implementing Quad Damage powerup but having to change C++ slighly and UE4 4.22 is not hot reloading so it is not allowing me to get access to some variables which are required.